refactor: relocate components
