+++ app/loading.tsx
@@ -1,9 +1,9 @@
-import Elipsis from '@/components/Ellipsis';
+import Ellipsis from '@/components/UI/Ellipsis';
 
 export default function Loading() {
   return (
     <div className="flex h-[calc(100vh-6rem)] w-screen items-center justify-center">
-      <Elipsis />
+      <Ellipsis />
     </div>
   );
 }
