+++ components/Response/ResponseBody.tsx
@@ -3,7 +3,7 @@ import { selectResponse } from '@/redux/features/mainSlice';
 import { useAppSelector } from '@/redux/hooks';
 import Editor from '@monaco-editor/react';
 import { Box, Paper, Table, TableBody, TableCell, TableContainer, TableHead, TableRow } from '@mui/material';
-import CustomTabPanel from '../Table/CustomTabPanel';
+import CustomTabPanel from '../UI/CustomTabPanel';
 
 type PropTypes = {
   activeTab: number;
